+++ FinancialMachineLearning/sample_weights/concurrency.py
@@ -7,12 +7,12 @@ def concurrent_events(close_series_index, label_endtime, molecule):
     label_endtime = label_endtime.loc[:label_endtime[molecule].max()]
 
     nearest_index = close_series_index.searchsorted(np.array([label_endtime.index[0], label_endtime.max()]))
-    count = pd.Series(0, index=close_series_index[nearest_index[0]:nearest_index[1] + 1])
+    count = pd.Series(0, index = close_series_index[nearest_index[0]:nearest_index[1] + 1])
     for t_in, t_out in label_endtime.items():
         count.loc[t_in:t_out] += 1
     return count.loc[molecule[0]:label_endtime[molecule].max()]
 def average_uniqueness(label_endtime, num_conc_events, molecule):
-    wght = pd.Series(index=molecule)
+    wght = pd.Series(dtype = float, index = molecule)
     for t_in, t_out in label_endtime.loc[wght.index].items():
         wght.loc[t_in] = (1. / num_conc_events.loc[t_in:t_out]).mean()
     return wght
